Update to allow for use of systemd and onedir based Tiamat salt-minion

+++ svtminion.sh
@@ -15,7 +15,7 @@ set -o pipefail
 # using bash for now
 # run this script as root, as needed to run salt
 
-SCRIPT_VERSION='2021.09.07.01'
+SCRIPT_VERSION='2021.09.08.02'
 
 # definitions
 
@@ -24,18 +24,28 @@ SCRIPT_VERSION='2021.09.07.01'
 
 ## Repository locations and naming
 readonly salt_name="salt"
-readonly salt_pkg_name="salt-3003-3-linux-amd64.tar.gz"
-readonly base_url="https://repo.saltproject.io/salt/singlebin"
-readonly salt_url="${base_url}/3003/${salt_pkg_name}"
-readonly salt_url_chksum="${base_url}/3003/salt-3003_SHA3_512"
+readonly salt_url_version="3003.3-1"
+readonly salt_pkg_name="${salt_name}-${salt_url_version}-linux-amd64.tar.gz"
+## readonly base_url="https://repo.saltproject.io/salt/singlebin"
+readonly base_url="https://repo.saltproject.io/salt/vmware-tools-onedir"
+readonly salt_url="${base_url}/${salt_url_version}/${salt_pkg_name}"
+readonly salt_url_chksum="${base_url}/${salt_url_version}/${salt_name}-${salt_url_version}_SHA_512"
 
 # Salt file and directory locations
 readonly base_salt_location="/opt/saltstack"
-readonly salt_dir="${base_salt_location}/salt"
-readonly test_exists_file="${salt_dir}/${salt_name}"
+readonly salt_dir="${base_salt_location}/${salt_name}"
+readonly test_exists_file="${salt_dir}/run/run"
+
 readonly salt_conf_dir="/etc/salt"
 readonly salt_minion_conf_name="minion"
 readonly salt_minion_conf_file="${salt_conf_dir}/${salt_minion_conf_name}"
+
+readonly salt_usr_bin_file_list="salt-minion
+salt-call
+"
+
+readonly salt_systemd_file_list="salt-minion.service
+"
 
 readonly list_file_dirs_to_remove="${base_salt_location}
 /etc/salt
@@ -43,7 +53,9 @@ readonly list_file_dirs_to_remove="${base_salt_location}
 /var/cache/salt
 /var/log/salt
 /etc/init.d/salt*
-/usr/lib/systemd/system/salt*.service
+/usr/bin/salt-*
+/usr/lib/systemd/system/salt-minion.service
+/etc/systemd/system/salt-minion.service
 "
 
 ## VMware file and directory locations
@@ -134,7 +146,7 @@ esac
 #
  _usage() {
      echo ""
-     echo "usage: ${0}  [-c|--status] [-d|--debug] [-e|--depends]"
+     echo "usage: ${0}  [-c|--status] [-d|--debug] [-e|--depend]"
      echo "             [-h|--help] [-i|--install] [-r|--remove] [-v|--verbose]"
      echo ""
      echo "  -c, --status    return status for this script"
@@ -164,16 +176,22 @@ _cleanup() {
     # don't know the worth of setting the current status, since script died
     if [[ ${CURRENT_STATUS} -eq ${STATUS_CODES[${installing}]} ]]; then
         CURRENT_STATUS=${STATUS_CODES[${installFailed}]}
+    elif [[ ${CURRENT_STATUS} -eq ${STATUS_CODES[${installed}]} ]]; then
+        # normal case with exit 0, but double-check
+        svpid=$(_find_salt_pid)
+        if [[ -z ${svpid} || ! -f ${test_exists_file} ]]; then
+            CURRENT_STATUS=${STATUS_CODES[${installFailed}]}
+        fi
     elif [[ ${CURRENT_STATUS} -eq ${STATUS_CODES[${removing}]} ]]; then
         CURRENT_STATUS=${STATUS_CODES[${removeFailed}]}
-        sleep 5
         svpid=$(_find_salt_pid)
         if [[ -z ${svpid} ]]; then
-            if [[ ! -d ${base_salt_location} ]]; then
+            if [[ ! -f ${test_exists_file} ]]; then
                 CURRENT_STATUS=${STATUS_CODES[$notInstalled]}
             fi
         fi
     else
+        # assume not installed
         CURRENT_STATUS=${STATUS_CODES[${notInstalled}]}
     fi
 }
@@ -183,9 +201,8 @@ _cleanup() {
 trap _cleanup INT EXIT
 
 ## cheap trim relying on echo to convert tabs to spaces and all multiple spaces to a single space
-## also gets rid of \ ' "
 _trim() {
-    echo "$1"
+    echo $1
 }
 
 # work functions
@@ -264,16 +281,18 @@ _fetch_salt_minion() {
     # could check if alreasdy there but by always getting it
     # ensure we are not using stale versions
     local retn=0
+    local curdir=$(pwd)
     CURRENT_STATUS=${STATUS_CODES[${installing}]}
-    mkdir -p ${salt_dir}
-    cd ${salt_dir}
+    mkdir -p ${base_salt_location}
+    cd ${base_salt_location}
     curl -o "${salt_pkg_name}" -fsSL "${salt_url}"
     tar -xvzf ${salt_pkg_name}
-    if [[ ! -f ${salt_name} ]]; then
+    if [[ ! -f ${test_exists_file} ]]; then
         CURRENT_STATUS=${STATUS_CODES[${installFailed}]}
         retn=1
     fi
     CURRENT_STATUS=${STATUS_CODES[${installed}]}
+    cd ${curdir}
     return ${retn}
 }
 
@@ -288,8 +307,8 @@ _fetch_salt_minion() {
 #
 
 _find_salt_pid() {
-    # find the pid for salt if active
-    local salt_pid=$(ps -ef | grep -v 'grep' | grep salt | head -n 1 | awk -F " " '{print $2}')
+    # find the pid for salt-minion if active
+    local salt_pid=$(ps -ef | grep -v 'grep' | grep "${salt_name}\/run\/run minion" | head -n 1 | awk -F " " '{print $2}')
     echo ${salt_pid}
 }
 
@@ -311,23 +330,26 @@ _find_salt_pid() {
 #   CURRENT_STATUS updated
 #
 # Results:
-#   Exits with ${CURRENT_STATUS}
+#   Echos ${CURRENT_STATUS}
 #
 
 _status_fn() {
     # return status
+    local retn_status=${STATUS_CODES[${notInstalled}]}
     if [[ ${CURRENT_STATUS} -eq  ${STATUS_CODES[${installing}]}
         || ${CURRENT_STATUS} -eq ${STATUS_CODES[${installFailed}]}
         || ${CURRENT_STATUS} -eq ${STATUS_CODES[${removing}]}
         || ${CURRENT_STATUS} -eq ${STATUS_CODES[${removeFailed}]} ]]; then
-        return ${CURRENT_STATUS}
-    fi
-    if [[ -f "${test_exists_file}" ]]; then
+        retn_status=${CURRENT_STATUS}
+    elif [[ -f "${test_exists_file}" ]]; then
         CURRENT_STATUS=${STATUS_CODES[${installed}]}
+        retn_status=${CURRENT_STATUS}
     else
         CURRENT_STATUS=${STATUS_CODES[${notInstalled}]}
+        retn_status=${CURRENT_STATUS}
     fi
-    return ${CURRENT_STATUS}
+    echo ${retn_status}
+    return 0
 }
 
 
@@ -361,7 +383,7 @@ _deps_chk_fn() {
 #  _install_fn
 #
 #   Executes scripts to install Salt from Salt repository
-#   and start the salt-minion
+#       and start the salt-minion using systemd
 #
 # Side Effects:
 #   CURRENT_STATUS updated
@@ -372,6 +394,7 @@ _deps_chk_fn() {
 
 _install_fn () {
     # execute install of Salt minion
+    local curdir=$(pwd)
     _fetch_salt_minion
     local retn=$?
 
@@ -379,9 +402,25 @@ _install_fn () {
     _fetch_vmtools_salt_minion_conf
     retn=$((${retn}|$?))
     if [[ ${retn} -eq 0 && -f "${test_exists_file}" ]]; then
-        ## echo "_install_fn starting ${test_exists_file} minion in bg"
-        ## $(nohup ${salt_dir}/${salt_name} "minion" &)
-        retn=0
+        # copy helper script for /usr/bin
+        for idx in "${salt_usr_bin_file_list}"; do
+            cp -a ${idx} /usr/bin/
+        done
+
+        # install salt-minion systemd service script
+        for idx in "${salt_systemd_file_list}"; do
+            cp -a ${idx} /usr/lib/systemd/system/
+            cd /etc/systemd/system
+            ln -s /usr/lib/systemd/system/${idx} ${idx}
+            cd ${curdir}
+
+            # start the salt-minion using systemd
+            systemctl daemon-reload
+            retn=$((${retn}|$?))
+            local name_service=$(echo ${idx} | cut -d '.' -f 1)
+            systemctl restart ${name_service}
+            retn=$((${retn}|$?))
+        done
     fi
     return ${retn}
 }
@@ -394,7 +433,8 @@ _install_fn () {
 #
 
  _remove_installed_files_dirs() {
-    for idx in "${list_file_dirs_to_remove}"; do
+    for idx in ${list_file_dirs_to_remove}
+    do
         rm -fR "${idx}"
     done
 }
@@ -404,6 +444,7 @@ _install_fn () {
 #  _uninstall_fn
 #
 #   Executes scripts to uninstall Salt from system
+#       stopping the salt-minion using systemd
 #
 # Side Effects:
 #   CURRENT_STATUS updated
@@ -415,24 +456,36 @@ _install_fn () {
 _uninstall_fn () {
     # remove Salt minion
     local retn=0
-    if [[ -f "${test_exists_file}" ]]; then
+    if [[ ! -f "${test_exists_file}" ]]; then
         CURRENT_STATUS=${STATUS_CODES[${notInstalled}]}
         retn=1
-    fi
-    CURRENT_STATUS=${STATUS_CODES[${removing}]}
-    svpid=$(_find_salt_pid)
-    if [[ -n ${svpid} ]]; then
-        kill ${svpid}
-        ## given it a little time
-        sleep 5
-    fi
-    svpid=$(_find_salt_pid)
-    if [[ -n ${svpid} ]]; then
-        CURRENT_STATUS=${STATUS_CODES[$removeFailed]}
-        retn=1
     else
-        _remove_installed_files_dirs
-        CURRENT_STATUS=${STATUS_CODES[$notInstalled]}
+        CURRENT_STATUS=${STATUS_CODES[${removing}]}
+        svpid=$(_find_salt_pid)
+        if [[ -n ${svpid} ]]; then
+            # stop the active salt-minion using systemd
+            # and give it a little time to stop
+            systemctl stop salt-minion
+            retn=$((${retn}|$?))
+            sleep 5
+        fi
+
+        if [[ ${retn} -eq 0 ]]; then
+            svpid=$(_find_salt_pid)
+            if [[ -n ${svpid} ]]; then
+                kill ${svpid}
+                ## given it a little time
+                sleep 5
+            fi
+            svpid=$(_find_salt_pid)
+            if [[ -n ${svpid} ]]; then
+                CURRENT_STATUS=${STATUS_CODES[$removeFailed]}
+                retn=1
+            else
+                _remove_installed_files_dirs
+                CURRENT_STATUS=${STATUS_CODES[$notInstalled]}
+            fi
+        fi
     fi
     return ${retn}
 }
@@ -483,7 +536,7 @@ while true; do
   case "$1" in
     -c | --status ) STATUS_CHK=1; shift ;;
     -d | --debug )  DEBUG_FLAG=1; shift ;;
-    -e | --depends ) DEPS_CHK=1; shift ;;
+    -e | --depend ) DEPS_CHK=1; shift ;;
     -h | --help ) USAGE_HELP=1; shift ;;
     -i | --install ) INSTALL_FLAG=1; shift ;;
     -r | --remove ) UNINSTALL_FLAG=1; shift ;;
@@ -502,54 +555,29 @@ fi
 
 ##  MAIN BODY OF SCRIPT
 
-_display "$SCRIPTNAME: autobuild started"
+## _display "$SCRIPTNAME: autobuild started"
 
 # check if salt-minion is installed
 if [[ -f "${test_exists_file}" ]]; then CURRENT_STATUS=${STATUS_CODES[$notInstalled]}; fi
-_ddebug "$SCRIPTNAME: CURRENT_STATUS on startup is ${CURRENT_STATUS}"
+## _ddebug "$SCRIPTNAME: CURRENT_STATUS on startup is ${CURRENT_STATUS}"
 
 retn=0
 
 if [[ ${STATUS_CHK} -eq 1 ]]; then
-    _status_fn
+    cur_status=$(_status_fn)
+    echo ${cur_status}
     retn=$?
-##    exit ${retn}
 elif [[ ${DEPS_CHK} -eq 1 ]]; then
     _deps_chk_fn
     retn=$?
-##    exit ${retn}
 elif [[ ${INSTALL_FLAG} -eq 1 ]]; then
     _install_fn
     retn=$?
-##    exit ${retn}
 elif [[ ${UNINSTALL_FLAG} -eq 1 ]]; then
     _uninstall_fn
     retn=$?
-##    exit ${retn}
 else
     _usage
 fi
-
-# doing this until onedir and daemonization is available
-# exit is by Cntl-C
-echo "before loop CURRENT_STATUS on startup is ${CURRENT_STATUS}"
-
-if [[ ${CURRENT_STATUS} -eq ${STATUS_CODES[${installed}]} ]]; then
-    echo "main loop starting ${salt_dir}/${salt_name} minion in bg"
-    $(nohup ${salt_dir}/${salt_name} "minion" &)
-fi
-while [[ ${CURRENT_STATUS} -eq ${STATUS_CODES[${installed}]} ]];
-do
-    echo "loop CURRENT_STATUS is '${CURRENT_STATUS}', code '${STATUS_CODES[${installed}]}'"
-    sleep 5
-    ## check
-    svpid=$(_find_salt_pid)
-    if [[ -z ${svpid} ]]; then
-        CURRENT_STATUS=${STATUS_CODES[${removeFailed}]}
-        if [[ ! -d ${base_salt_location} ]]; then
-            CURRENT_STATUS=${STATUS_CODES[$notInstalled]}
-        fi
-    fi
-done
 
 exit ${retn}
